refactor(agents): isolate durable yielded requester recovery
